Refactor add comment
