+++ src/modules/addComment.js
@@ -1,3 +1,6 @@
+import { postComments, getCommentsData } from './getComments.js';
+import commentTemplate from './addComment.js';
+
 const commentTemplate = (data, container) => {
     data.forEach((item => {
         const comment = document.createElement('div');
@@ -10,4 +13,21 @@ const commentTemplate = (data, container) => {
     }));
 }
 
-export default commentTemplate;
+const addComment = () => {
+document.addEventListener('click', async (e) => {
+    e.preventDefault();
+    const commentContainer = e.target.parentNode.parentNode.parentNode.childNodes[1].childNodes[3];
+    const comments = e.target.previousElementSibling;
+    const title = e.target.parentNode.childNodes[3];
+    const showId = Number(e.target.parentNode.parentNode.parentNode.parentNode.id);
+    if (e.target.id === 'submit') {
+        await postComments(showId, title.value, comments.value);
+        const commentsData = await getCommentsData(showId);
+        commentTemplate(commentsData, commentContainer);
+    }
+});
+}
+
+export default addComment;
+
+
